requirements.txt
torch>=1.10.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
faiss-cpu>=1.7.2
matplotlib>=3.5.0
tqdm>=4.62.0